# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  comment.Comment:
    properties:
      content:
        example: Hello, world!
        type: string
      created_at:
        example: "2021-09-01T12:00:00Z"
        type: string
      id:
        example: 3
        type: integer
      is_deleted:
        example: false
        type: boolean
      parent_id:
        example: 1
        type: integer
      reply_id:
        example: 3
        type: integer
      user:
        $ref: '#/definitions/comment.User'
    required:
    - content
    - created_at
    - id
    - is_deleted
    - user
    type: object
  comment.Create:
    properties:
      content:
        example: Hello, world!
        type: string
      parent_id:
        example: 1
        type: integer
      reply_id:
        example: 3
        type: integer
    required:
    - content
    type: object
  comment.GetAllCommentsResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/pagination.Pagination'
      status:
        $ref: '#/definitions/model.status'
    type: object
  comment.Update:
    properties:
      content:
        example: Hello, world!
        type: string
    required:
    - content
    type: object
  comment.User:
    properties:
      id:
        example: 1
        type: integer
      username:
        example: Jack_Vorobey123
        type: string
    type: object
  model.Response:
    properties:
      data: {}
      status:
        $ref: '#/definitions/model.status'
    type: object
  model.status:
    enum:
    - success
    - error
    type: string
    x-enum-varnames:
    - success
    - err
  pagination.Pagination:
    properties:
      current_page:
        example: 1
        type: integer
      per_page:
        example: 1
        type: integer
      total:
        example: 10
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for the SOS Kotopes project. It provides endpoints
    for managing the database of animals in need of help.
  title: SOS Kotopes API
  version: "1.0"
paths:
...
  /posts/{post_id}/comments:
    get:
      consumes:
      - application/json
      description: Get all comments for a post
      operationId: get-all-comments
      parameters:
      - description: Post ID
        in: path
        minimum: 1
        name: post_id
        required: true
        type: integer
      - description: Limit
        in: query
        minimum: 1
        name: limit
        required: true
        type: integer
      - description: Offset
        in: query
        minimum: 0
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/comment.GetAllCommentsResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/comment.Comment'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get all comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Create a comment for a post
      operationId: create-comment
      parameters:
      - description: Post ID
        in: path
        minimum: 1
        name: post_id
        required: true
        type: integer
      - description: Comment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/comment.Create'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/comment.Comment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Create a comment
      tags:
      - comments
  /posts/{post_id}/comments/{comment_id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment for a post
      operationId: delete-comment
      parameters:
      - description: Post ID
        in: path
        minimum: 1
        name: post_id
        required: true
        type: integer
      - description: Comment ID
        in: path
        minimum: 1
        name: comment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Delete a comment
      tags:
      - comments
    patch:
      consumes:
      - application/json
      description: Update a comment for a post
      operationId: update-comment
      parameters:
      - description: Post ID
        in: path
        minimum: 1
        name: post_id
        required: true
        type: integer
      - description: Comment ID
        in: path
        minimum: 1
        name: comment_id
        required: true
        type: integer
      - description: Comment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/comment.Update'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/comment.Comment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Update a comment
      tags:
      - comments
...
securityDefinitions:
  ApiKeyAuthBasic:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"